requirements.txt
pandas==1.3.4
Pillow==9.2.0
numpy==1.20.3
chardet==4.0.0
scikit-learn==1.1.1
torchvision==0.13.0
torch==1.12.0
transformers==4.19.2
sklearn==0.0
